Solution to Convert Decimal to Binary

+++ Convert Decimal to Binary@@67882a731b4392d8fc1577f5/code.c
@@ -3,7 +3,7 @@
 
 int main()
 {
-    int n;
+    long int n;
     scanf("%d", &n);
     int answer = 0;
     int i = 1;
